Move selection to own module
