+++ src/org/sosy_lab/cpachecker/core/algorithm/legion/LegionAlgorithm.java
@@ -33,6 +33,8 @@ import org.sosy_lab.common.configuration.Option;
 import org.sosy_lab.common.configuration.Options;
 import org.sosy_lab.common.log.LogManager;
 import org.sosy_lab.cpachecker.core.algorithm.Algorithm;
+import org.sosy_lab.cpachecker.core.algorithm.legion.selection.RandomSelectionStrategy;
+import org.sosy_lab.cpachecker.core.algorithm.legion.selection.Selector;
 import org.sosy_lab.cpachecker.core.interfaces.ConfigurableProgramAnalysis;
 import org.sosy_lab.cpachecker.core.interfaces.Property;
 import org.sosy_lab.cpachecker.core.reachedset.ReachedSet;
