+++ src/org/sosy_lab/cpachecker/core/algorithm/legion/selection/Selector.java
@@ -1,9 +1,9 @@
-package org.sosy_lab.cpachecker.core.algorithm.legion;
+package org.sosy_lab.cpachecker.core.algorithm.legion.selection;
 
 import org.sosy_lab.cpachecker.core.reachedset.ReachedSet;
 import org.sosy_lab.java_smt.api.BooleanFormula;
 
-interface Selector {
+public interface Selector {
     /**
      * Select a target state and return the path formula that leads to it.
      */
